Sync with the private repository
RENDER_FLUSH_PRE/RENDER_FLUSH_POST -> RENDER_PRE/RENDER_POST

Merge the desc data.
If it is not read by the script engine of provider,
merge new desc blocks with previous one.

script engine should know the context of desc blocks.

Accessibility desc type added

Add dummy implmentation files.

These patch will not apply any changes to release binary.
Just keep the implementations only in this repository as source code level.
will not be built and distributed.
---------------------------------

Implementing the snapshot & virtual window add function.

The snapshot window is used for getting an image file from the rendering buffer.

Conceptual logic (how it works?)
-----------------------------------------------------------------------
1. Create a new snapshot window

2. Add widget objects for representing the content.

3. Trigger the snapshot flush opertaion

4. post-render callback will add the flush timer (0.5 sec period)

5. start rendering loop

5.1 If the pre render callback is called,
  reset the flush timer

5.2 If the flush timer is expired so the timer callback is called,
  flush the data in the rendering buffer to a file.
-----------------------------------------------------------------------
This logic can be failed when the system is going to slower and slower.
After set the flush timer from the post render callback, the process
has to start the rendering of a next frame in 0.5 sec.
If it couldn't start rendering, the flush timer will be expired.
then it will finsish the rendering loop and flush the rendered data to a file
If there are remained frames, we will lost it.

Change-Id: Ic7216ce392faaf679f8f024130599516477ffb2e